🔥 feat: Add new components for Academic Features, Contact, Gallery, Hero Section, and Notice Events
- Implemented AcademicFeaturesSection for showcasing academic offerings.
- Created ContactPage component for displaying contact information and admission details.
- Developed GallerySection to present a photo gallery of campus activities.
- Introduced HeroSection for a welcoming banner with image loading handling.
- Added EventCard and NoticeCard components for displaying events and notices.
- Refactored NoticeEventsSection to utilize new NoticeCard and EventCard components.
- Removed unused TestComponents.js file.
