+++ components/Gallery/GallerySection.jsx
